requirements.txt
attrs==22.1.0
bottle==0.12.23
exceptiongroup==1.0.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
pyparsing==3.0.9
PyYAML==6.0
tomli==2.0.1
tornado==6.2
